Do not automatically create a restic repository
Although automatically creating a restic repository was convenient, it
could lead to accidental data loss if the wrong directory was used as
the argument. It is safer to require the user to manually create a
repository first and for the script to exit with a non-zero status code
if a repository is not detected.