Handle repositories containing spaces in tar wrapper
The tar wrapper script has the following logic:

1. Use find to get all the existing tar files
2. Sort the files in reverse order
3. Create an array of the files
4. Iterate over the files and delete any after the first 5

However, this failed when the path contained spaces, because Bash uses
spaces to create each element in the array. Messing around with IFS,
read, find -print0 etc. failed to fix the problem.

Thinking about this further, I realised that the only part of the path
that could contain spaces was the REPO variable, as that is supplied by
the user. So the spaces can be eliminated by:

1. Changing into REPO
2. Run file using the current directory
3. No spaces are present so Bash can convert the output to an array
4. Iterate over array and delete files
4. Change back into the previous directory

This is a bit of a hack, but it works.

Closes #13